fix some warnings

+++ Eclipse/src/Editor/Windows/Profiler/Profiler.cpp
@@ -127,7 +127,7 @@ namespace Eclipse
 	void ProfilerWindow::PrintCpuPercentage(float value)
 	{
 
-		ImGui::TextColored(ImVec4(0.9f, 0.6f, 0.1f, 1.0f), "CUP USAGE :%.2f %%", (((value) / ProfilerWindow::engine_time) * 100));
+		ImGui::TextColored(ImVec4(0.9f, 0.6f, 0.1f, 1.0f), "CUP USAGE :%.2f %%", static_cast<float>(((value / ProfilerWindow::engine_time) * 100.f)));
 	}
 }
